Refactor of TouchID and iOS Keychain Re-Login code. Now we can store multiple username/password combos in the keychain so multiple accounts can auth with TouchID.
Because of this, we now have to ask the user if they want to enable TouchID. This is to prevent logging into a friend's phone from giving the friend access to the account after logout. TouchID will now require the user's password to enable. If they logged in w/password, they won't need it to enable.