Dismiss any open Fading Alert dialogues when entering SendViewController.
Since we might enter this view upon login, the Loading Transactions alert might be up due to the TransactionsViewController popping it up first before we get bounced to SendView. SendView manages it's own set of popups depending on the current loading state of wallets. Since we use SendView for BitID login, we need it to be usable instantly upon login which is possible since the BitID private keys are instantly decrypted from the account before transactions are loaded.